+++ regular.h
@@ -21,10 +21,10 @@ public:
     ~Regular();
 
 private:
-    QPixmap *regularPixmap;
-    int xPos;
-    int yPos;
-    void move();
+    QPixmap *regularPixmap;     // Holds image.
+    double xPos;                   // Holds x-position.
+    double yPos;                   // Holds y-position.
+    void move();                // Moves zombie until collision with plants.
 
     // Virtual derived functions.
     void paint(QPainter *painter, const QStyleOptionGraphicsItem *, QWidget *);
